LayoutTests/http/tests/security/mixedContent/insecure-localhost-image-in-main-frame-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-localhost-image.html requested insecure content from http://localhost:8080/security/resources/compass.jpg. This content was not upgraded and must be served from the local host.

This test opens a window that loads an insecure image from localhost. We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
